dynamic-not-editable-pricing/requirements.txt
python-dotenv
polar-sdk
fastapi[standard]
